+++ itests/app_1.py
@@ -17,6 +17,7 @@ from blacksheep import (
     json,
     text,
 )
+from blacksheep.contents import ASGIContent
 from blacksheep.server.compression import use_gzip_compression
 from itests.utils import CrashTest, ensure_folder
 
@@ -240,6 +241,38 @@ async def send_file_with_bytes_io():
         file_name="data.txt",
         content_disposition=ContentDispositionType.INLINE,
     )
+
+
+@app.router.get("/check-disconnected")
+async def check_disconnected(request: Request, expect_disconnected: bool):
+    check_file = pathlib.Path(".is-disconnected.txt")
+    assert await request.is_disconnected() is False
+    # Simulate a delay
+    await asyncio.sleep(0.3)
+
+    if expect_disconnected:
+        # Testing the scenario when the client disconnected
+        assert (
+            await request.is_disconnected()
+        ), "The client disconnected and this should be visible"
+        check_file.write_text("The connection was disconnected")
+    else:
+        assert (
+            await request.is_disconnected() is False
+        ), "The client did not disconnect and this should be visible"
+
+    return "OK"
+
+
+@app.router.get("/read-asgi-receive")
+def check_asgi_receive_readable(request: Request):
+    content = request.content
+    assert isinstance(content, ASGIContent)
+
+    receive = content.receive
+    assert callable(receive)
+
+    return "OK"
 
 
 if __name__ == "__main__":