+++ itests/utils.py
@@ -1,6 +1,8 @@
 import errno
 import os
 import socket
+from contextlib import contextmanager
+from pathlib import Path
 from urllib.parse import urljoin
 
 import requests
@@ -75,3 +77,15 @@ def get_sleep_time():
 
 def get_test_files_url(url: str):
     return f"my-cdn-foo:{url}"
+
+
+@contextmanager
+def temp_file(name: str):
+    file = Path(name)
+    if file.exists():
+        file.unlink()
+
+    yield file
+
+    if file.exists():
+        file.unlink()